refactor(canon): keep virtual_ts files within source-length budget
The test:scripts source-file-length check failed because the prior commit
grew two already-over-limit files past their allowed ceilings:
virtual_ts/tests.rs (2298 > 2268) and virtual_ts/scope/closures.rs
(491 > 486).

Split by ownership boundary instead of bumping the ratchet:
- Move the new legacy Nuxt 2 page test into a dedicated
  tests/legacy_nuxt2_page_context.rs submodule.
- Compute the listener rest-args type (loose Vue 2 wrapper vs resolved
  emit tuple) in generate_component_event_types, which already branches on
  legacy_vue2 and is under the limit, and consume it via a single append!
  in closures.rs.

No behavior change; the generated virtual TS is identical.
