+++ crates/vize_canon/src/virtual_ts/scope/closures.rs
@@ -370,23 +370,17 @@ fn generate_scope_node(
                 .expect("component event handler should have a target component");
                 let event_type = event_types.event_type;
                 let args_type = event_types.args_type;
+                let listener_args_type = event_types.listener_args_type;
                 let listener_type = event_types.listener_type;
                 // Type the listener against the FULL emit argument tuple so
                 // multi-arg emits keep every parameter (#1512). When the emit
                 // signature stays unresolved (`unknown[]`, e.g. a fallthrough
                 // DOM event on a component), fall back to the single `$event`
                 // parameter so those handlers keep type-checking.
-                if ctx.legacy_vue2 {
-                    append!(
-                        *ts,
-                        "{indent}type {listener_type} = unknown[] extends {args_type} ? (($event: {event_type}) => unknown) : ((...args: __VizeVue2LooseEmitArgs<{args_type}>) => unknown);\n",
-                    );
-                } else {
-                    append!(
-                        *ts,
-                        "{indent}type {listener_type} = unknown[] extends {args_type} ? (($event: {event_type}) => unknown) : ((...args: {args_type}) => unknown);\n",
-                    );
-                }
+                append!(
+                    *ts,
+                    "{indent}type {listener_type} = unknown[] extends {args_type} ? (($event: {event_type}) => unknown) : ((...args: {listener_args_type}) => unknown);\n",
+                );
                 // Receive every listener argument via a rest parameter typed by
                 // `Parameters<listener>` (always a tuple, so the spread targets a
                 // rest parameter and avoids TS2556). `$event` stays bound to the
